+++ packages/engine/src/frontmatter/yaml.ts
@@ -37,7 +37,7 @@ const splitYamlPair = (
       1,
     );
   }
-  return [substringCount(text, 0, separator).trim(), substringFrom(text, separator + 1).trim()];
+  return [substringCount(text, 0, separator as int32).trim(), substringFrom(text, (separator + 1) as int32).trim()];
 };
 
 const applyMenuProperty = (
